requirements.txt
temporalio>=1.4.0
python-dotenv>=1.0.0 